requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.61.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.5
